requirements.txt
apache-airflow-providers-mysql
apache-airflow-providers-postgres
pandas
mysqlclient
psycopg2-binary